backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
onnxruntime-gpu>=1.16.0  # ONNX Runtime GPU 支持(Inpaint 模型)
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8
realesrgan==0.3.0
